pyproject.toml
[project]
name = "ring-central-zoho-desk-bridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "botbuilder-core>=4.17.1",
    "botbuilder-integration-aiohttp>=4.17.1",
    "aiofiles>=23.0",
]
